This is the initial commit for the core 3.0.0 build. Architectural changes have been completed and the project will compile although the functionality is not complete.
Several new concepts have been added to this build, specifically the use of ServerAgents, which include the traditional logging provider, but now include performance, session and storage agents. In addition the old configuration system have been removed and replaced with a object orientated version based around the content entities with well defined schema based configuration files.

Please note: this code is till in development and is not a final release.